+++ src/cosmos/cosmosInfos.ts
@@ -1,6 +1,7 @@
 import { axelar } from './info/axelarInfo'
 import { coreum } from './info/coreumInfo'
 import { cosmoshub } from './info/cosmoshubInfo'
+import { nym } from './info/nymInfo'
 import { osmosis } from './info/osmosisInfo'
 import { thorchainrune } from './info/thorchainruneInfo'
 import { thorchainrunestagenet } from './info/thorchainruneStagenetInfo'
@@ -9,6 +10,7 @@ export const cosmosPlugins = {
   axelar,
   coreum,
   cosmoshub,
+  nym,
   osmosis,
   thorchainrune,
   thorchainrunestagenet