+++ src/ethereum/ethereumInfos.ts
@@ -17,6 +17,7 @@ import { filecoinfevm } from './info/filecoinFevmInfo'
 import { holesky } from './info/holeskyInfo'
 import { hyperevm } from './info/hyperEvmInfo'
 import { monad } from './info/monadInfo'
+import { opbnb } from './info/opbnbInfo'
 import { optimism } from './info/optimismInfo'
 import { polygon } from './info/polygonInfo'
 import { pulsechain } from './info/pulsechainInfo'
@@ -45,6 +46,7 @@ export const ethereumPlugins = {
   holesky,
   hyperevm,
   monad,
+  opbnb,
   optimism,
   polygon,
   pulsechain,