+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Add Monad (MON) support
+
 ## 4.71.2 (2026-02-02)
 
 - fixed: Remove extra call to whitelistAssets that was hanging and preventing core callbacks from resolving