+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: Add Monad (MON) support
+- added: Nym (NYM) Cosmos-based chain with NYX builtin token
 
 ## 4.71.2 (2026-02-02)